+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * Fix incorrect URL in PrecisionRecallController docstring
+* Add scientific references for regression conformity scores
 * Fix double inference when using `predict_set` function in split conformal classification
 * Add FAQ entry in the documentation about ongoing works to extend MAPIE for LLM control
 * MAPIE now supports Python versions up to the latest release (currently 3.13)